fix: rename forward icon to share icon (#1099)
It was unclear which one was the ShareIcon
Renaming the ForwardIcon to ShareIcon

WT-253 #done
