+++ packages/extension/src/companion/CompanionMenu.tsx
@@ -5,7 +5,7 @@ import CommentIcon from '@dailydotdev/shared/src/components/icons/Discuss';
 import BookmarkIcon from '@dailydotdev/shared/src/components/icons/Bookmark';
 import MenuIcon from '@dailydotdev/shared/src/components/icons/Menu';
 import ArrowIcon from '@dailydotdev/shared/src/components/icons/Arrow';
-import ForwardIcon from '@dailydotdev/shared/src/components/icons/Forward';
+import ShareIcon from '@dailydotdev/shared/src/components/icons/Share';
 import LogoIcon from '@dailydotdev/shared/src/svg/LogoIcon';
 import classNames from 'classnames';
 import SimpleTooltip from '@dailydotdev/shared/src/components/tooltips/SimpleTooltip';
@@ -268,7 +268,7 @@ export default function CompanionMenu({
             buttonSize="medium"
             className="btn-tertiary-cabbage"
             onClick={onShare}
-            icon={<ForwardIcon />}
+            icon={<ShareIcon />}
           />
         </SimpleTooltip>
       )}
