+++ packages/shared/src/components/PostOptionsReadingHistoryMenu.tsx
@@ -3,7 +3,7 @@ import { Item } from '@dailydotdev/react-contexify';
 import dynamic from 'next/dynamic';
 import { QueryClient, QueryKey, useQueryClient } from 'react-query';
 import { ReadHistoryPost } from '../graphql/posts';
-import ShareIcon from './icons/Forward';
+import ShareIcon from './icons/Share';
 import BookmarkIcon from './icons/Bookmark';
 import XIcon from './icons/Close';
 import useBookmarkPost from '../hooks/useBookmarkPost';
